web: try to display the RSS feed entries on one line each.
Change the date from "Sunday, September 2013" to "07 September"; it's a
shorter, yet correct, date format that carries everything one expects.
Keeping the year is of little use since it's very uncommon to have entries
from the previous year: it's not ambiguous.

The date was also always on a separate line; make that non-forced.