intrinsics: don't include {d,f,i}vec.h from inside intrin.h.
Prior to this commit, #include <intrin.h> would automatically pull these
*vec.h headers for C++. Depending on compiler options (architecture and
optimizations), it would fail because they use functions which are only
declared later on in intrin.h.

Simply remove the inclusion of {d,f,i}vec.h from intrin.h since there is
little reason to do so and other compilers don't seem to do it.

This changes behaviour with previous mingw-w64 releases but only to make
it more compatible with MSVC. I originally investigated the issue to fix
a build of Qt 5.2.