web: rm mibbit.html which doesn't work anymore.
Mibbit has been banned on many IRC networks beccause it makes it
difficult to ban people and makes it easy to circumvent bans. It hasn't
worked for OFTC for quite a long time and the link to mibbit.html has
been removed months ago from the website. This commit merely removes the
leftover file.

Note that OFTC has a its own webchat own but I haven't checked it
(yet?).