web: display at most 30 related projects but always show friend projects.
The list of related projects has been getting fairly long, almost a
hundred. This splits the list into two: friend projects (wine, ros, cygwin)
and every other. The others are again split into categories: package
providers, compilers/IDEs/tools and users.

The friend projects are always shown on the main page. For the others, 30
random ones are shown (all categories merged).

There is also a new webpage with the full list. It's currently a bit bare
but the idea is to have some more space to write descriptions and so on.